Move topicConnection components to shared
